requirements.txt
obs-websocket-py>=1.0.0
asyncio
typing-extensions>=4.0.0
python-dotenv>=0.19.0